Delete album and images

+++ pixsalle-template/src/Controller/PortfolioController.php
@@ -91,7 +91,7 @@ class PortfolioController
 		$data = $request->getParsedBody();
 		$imageURL = $data['imageId'];
 
-		if (isset($imageURL)) {
+		if (isset($imageURL) && $imageURL != "") {
 			$this->portfolioRepository->deletePhotoFromAlbum($id, $_SESSION['user_id'], $imageURL);
 		} else {
 			$this->portfolioRepository->deleteAlbum($id, $_SESSION['user_id']);
